DEV: Add `null` keyword to show targets when placeholder is present (#6)
When working with dropdowns, there are times when we want to _display_ the Target User Fields until a proper value is selected and then kick in the normal validations. 

Eg
<img width="268" alt="Screenshot 2024-01-05 at 10 18 45 AM" src="https://github.com/discourse/discourse-authentication-validations/assets/50783505/5e9d7910-7de6-4033-94c1-d308fa7d50a9">

To display the Target User Fields until a proper value is selected
- create a dropdown User Field
- check `has_custom_validation` on a User Field
- add `null` to show_values

and it will display any specified targets until a proper value has been selected. It will then go about with normal validation checking.